+++ src/controllers/api/relations/block.ts
@@ -1,5 +1,5 @@
 import express from "express";
-import getFirst from "../../../utils/getFirst";
+import getFirst from "../../../util/getFirst";
 import type { Relation } from "../../../types/relation";
 
 // For blocking users
