+++ src/controllers/api/relations/follow.ts
@@ -1,7 +1,7 @@
 import express from "express";
-import getFirst from "../../../utils/getFirst";
+import getFirst from "../../../util/getFirst";
 import type { Relation } from "../../../types/relation";
-import checkStatus from "../../../utils/checkStatus";
+import checkStatus from "../../../util/checkStatus";
 
 // For following another user
 const follow = async (req: express.Request, res: express.Response) => {
