+++ src/controllers/api/relations/following.ts
@@ -1,6 +1,6 @@
 import express from "express";
 import pg from "../../../db/pg";
-import checkStatus from "../../../utils/checkStatus";
+import checkStatus from "../../../util/checkStatus";
 
 // For getting the people whom the account follows
 const following = async (req: express.Request, res: express.Response) => {
