+++ src/controllers/auth/verify.ts
@@ -1,8 +1,8 @@
 import bcrypt from "bcrypt";
 import express from "express";
 import redis from "../../db/redis";
-import getFirst from "../../utils/getFirst";
-import { createJwt } from "../../utils/jwt";
+import getFirst from "../../util/getFirst";
+import { createJwt } from "../../util/jwt";
 import type { User } from "../../types/user";
 
 const verify = (req: express.Request, res: express.Response) => {
