+++ src/util/checkStatus.ts
