+++ src/util/getFirst.ts
